requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==24.3.0
audioop-lts==0.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
distro==1.9.0
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.1.0
openai==1.86.0
-e git+https://github.com/iamllcoolray/private_dan_bot.git@42318cb4731d52b6180459c7bddd511334746313#egg=private_dan_bot
propcache==0.2.1
py-cord==2.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.3.0
yarl==1.18.3
